requirements.txt
kafka-python==2.0.2
elasticsearch>=7.13.0,<9.0.0
faker==20.1.0
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.0.3
pyyaml==6.0
python-dotenv==1.0.0
